Fixed Some Bugs in all Versions

+++ src/V4/mnist.cu
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 
 // Read MNIST dataset
-double* loadMNISTImages(const char* filename, int numImages) {
+float* loadMNISTImages(const char* filename, int numImages) {
     if (VERBOSE) printf("Loading MNIST images from %s...\n", filename);
     FILE* file = fopen(filename, "rb");
     if (!file) {
@@ -12,7 +12,7 @@ double* loadMNISTImages(const char* filename, int numImages) {
         exit(1);
     }
     fseek(file, 16, SEEK_SET);
-    double* images = allocateMatrix(numImages, INPUT_SIZE);
+    float* images = allocateMatrix(numImages, INPUT_SIZE);
     for (int i = 0; i < numImages; i++) {
         for (int j = 0; j < INPUT_SIZE; j++) {
             unsigned char pixel;
@@ -23,7 +23,7 @@ double* loadMNISTImages(const char* filename, int numImages) {
                 exit(EXIT_FAILURE);
             }
 
-            images[i*INPUT_SIZE + j] = pixel / 255.0;
+            images[i*INPUT_SIZE + j] = pixel / 255.0f;
         }
         if (VERBOSE && i % 10000 == 0) printf("Loaded %d images\n", i);
     }
@@ -37,7 +37,7 @@ double* loadMNISTImages(const char* filename, int numImages) {
     return images;
 }
 
-double* loadMNISTLabels(const char* filename, int numLabels) {
+float* loadMNISTLabels(const char* filename, int numLabels) {
     if (VERBOSE) printf("Loading MNIST labels from %s...\n", filename);
     FILE* file = fopen(filename, "rb");
     if (!file) {
@@ -45,7 +45,7 @@ double* loadMNISTLabels(const char* filename, int numLabels) {
         exit(1);
     }
     fseek(file, 8, SEEK_SET);
-    double* labels = allocateMatrix(numLabels, OUTPUT_SIZE);
+    float* labels = allocateMatrix(numLabels, OUTPUT_SIZE);
     for (int i = 0; i < numLabels; i++) {
         unsigned char label;
         if (fread(&label, sizeof(unsigned char), 1, file) != 1) {
