+++ src/V4/mnist.h
@@ -4,7 +4,7 @@
 #include "nn.h"
 
 // MNIST data loading functions
-double* loadMNISTImages(const char* filename, int numImages);
-double* loadMNISTLabels(const char* filename, int numLabels);
+float* loadMNISTImages(const char* filename, int numImages);
+float* loadMNISTLabels(const char* filename, int numLabels);
 
 #endif
